CPR-1107 Split retry configuration
