+++ src/main/kotlin/uk/gov/justice/digital/hmpps/personrecord/client/CorePersonRecordAndDeliusClient.kt
@@ -2,13 +2,10 @@ package uk.gov.justice.digital.hmpps.personrecord.client
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties
 import org.springframework.data.web.PagedModel.PageMetadata
-import org.springframework.retry.annotation.Backoff
 import org.springframework.stereotype.Component
 import org.springframework.web.reactive.function.client.WebClient
-import org.springframework.web.reactive.function.client.WebClientResponseException
 import org.springframework.web.reactive.function.client.bodyToMono
 import reactor.core.publisher.Mono
-import uk.gov.justice.digital.hmpps.personrecord.CprRetryable
 import uk.gov.justice.digital.hmpps.personrecord.client.model.offender.ProbationAddress
 import uk.gov.justice.digital.hmpps.personrecord.client.model.offender.ProbationCase
 import uk.gov.justice.digital.hmpps.personrecord.model.person.Address
@@ -17,7 +14,7 @@ import uk.gov.justice.digital.hmpps.personrecord.model.person.Person
 @Component
 class CorePersonRecordAndDeliusClient(private val corePersonRecordAndDeliusWebClient: WebClient) {
 
-  @CprRetryable(retryFor = [WebClientResponseException.NotFound::class], backoff = Backoff(delay = 500, maxDelay = 1000), maxAttempts = "3")
+  @WebRetryable
   fun getPerson(crn: String): Person = Person.from(getProbationCase(crn))
 
   fun getProbationCase(crn: String): ProbationCase = fetchProbationCase(crn).block()!!
