+++ src/main/kotlin/uk/gov/justice/digital/hmpps/personrecord/client/WebRetryable.kt
@@ -1,29 +1,25 @@
-package uk.gov.justice.digital.hmpps.personrecord
+package uk.gov.justice.digital.hmpps.personrecord.client
 
-import jakarta.persistence.OptimisticLockException
 import org.springframework.core.annotation.AliasFor
-import org.springframework.dao.CannotAcquireLockException
-import org.springframework.dao.DataIntegrityViolationException
 import org.springframework.retry.annotation.Backoff
 import org.springframework.retry.annotation.Retryable
+import org.springframework.web.reactive.function.client.WebClientResponseException
 import kotlin.reflect.KClass
 
 @Target(AnnotationTarget.FUNCTION, AnnotationTarget.CLASS)
 @Retention(AnnotationRetention.RUNTIME)
 @MustBeDocumented
 @Retryable
-annotation class CprRetryable(
+annotation class WebRetryable(
 
   @get:AliasFor(annotation = Retryable::class, attribute = "maxAttemptsExpression")
-  val maxAttempts: String = "5",
+  val maxAttempts: String = "3",
 
   @get:AliasFor(annotation = Retryable::class, attribute = "retryFor")
   val retryFor: Array<KClass<out Throwable>> = [
-    OptimisticLockException::class,
-    DataIntegrityViolationException::class,
-    CannotAcquireLockException::class,
+    WebClientResponseException.NotFound::class
   ],
 
   @get:AliasFor(annotation = Retryable::class, attribute = "backoff")
-  val backoff: Backoff = Backoff(delay = 200, random = true, multiplier = 3.0),
+  val backoff: Backoff = Backoff(delay = 500, maxDelay = 1000),
 )
