+++ src/main/kotlin/uk/gov/justice/digital/hmpps/personrecord/service/queue/SQSMessageProcessor.kt
@@ -7,9 +7,9 @@ import org.springframework.web.reactive.function.client.WebClientResponseExcepti
 import reactor.core.publisher.Mono
 import tools.jackson.databind.json.JsonMapper
 import tools.jackson.module.kotlin.readValue
-import uk.gov.justice.digital.hmpps.personrecord.CprRetryable
 import uk.gov.justice.digital.hmpps.personrecord.client.model.sqs.NOTIFICATION
 import uk.gov.justice.digital.hmpps.personrecord.client.model.sqs.SQSMessage
+import uk.gov.justice.digital.hmpps.personrecord.jpa.DatabaseRetryable
 import uk.gov.justice.digital.hmpps.personrecord.service.TimeoutExecutor
 
 @Component
@@ -17,7 +17,7 @@ class SQSMessageProcessor(
   private val jsonMapper: JsonMapper,
 ) {
 
-  @CprRetryable
+  @DatabaseRetryable
   fun process(rawMessage: String, action: (sqsMessage: SQSMessage) -> Unit) = TimeoutExecutor.runWithTimeout {
     val sqsMessage = jsonMapper.readValue<SQSMessage>(rawMessage)
     try {
